Lab04.1/test/com/entertainment/catalog/CatalogTest.java
/*
 * This code is sample code, provided as-is, and we make no
 * warranties as to its correctness or suitability for
 * any purpose.
 *
 * We hope that it's useful to you.  Enjoy.
 * Copyright LearningPatterns Inc.
 */
package com.entertainment.catalog;

import static org.junit.Assert.*;
import java.util.Collection;
import java.util.Collections;
import org.junit.Test;
import com.entertainment.Television;

public class CatalogTest {
  
  /**
   * Contract: a no-matches result should be an empty collection (not null).
   */
  @Test
  public void testFindByBrandNoMatches() {
    Collection<Television> tvs = Catalog.findByBrand("NO-MATCHES");
    assertNotNull(tvs); //NOTE a pointer to an empty collection is different from null. Null means something isn't pointing to anything.
    assertTrue(tvs.isEmpty()); //NOTE asserts that the collection its pointing to is empty.
  }

  @Test
  public void testFindByBrandWithMatches(){
    Collection<Television> tvs = Catalog.findByBrand("Sony");
    assertNotNull(tvs);
    assertEquals(7, tvs.size());
    for(Television tv : tvs){
      assertEquals("Sony", tv.getBrand());
    }

  }

  @Test
  public void findByBrands() {
  }
}